Replace hashed passwords map with bloom filter
Storing multiple hashes for each password and the password itself on a
map can use very large amounts of memory when used with dictionary
files with a large number of entries.

Using a bloom filter allows us to reduce the amount memory used at the
expense of returning some false positives.

The user can choose a target a false-positive rate through the
HashFalsePositiveRate option which by default is 1%. But lowering the
rate increases memory consumption.

Another trade-off is that we lose the ability to know the original word
that caused the collision. Assuming most users are not interested in
that information and would generate another password anyways, this seems
reasonable trade-off.